feat(auth): silent refresh token rotation + Decimal type safety + PII masking
- auth/dependencies.py: _try_refresh() issues new access+refresh cookies on expiry
- auth/dependencies.py: _decode_access_token() rejects type=refresh used as access
- auth/service.py: create_refresh_token() with type='refresh' claim + 7-day expiry
- core/config.py: ACCESS_TOKEN_EXPIRE_MINUTES=15, REFRESH_TOKEN_EXPIRE_MINUTES=10080
- core/security.py: stripped to mask_sensitive_data only (auth moved to auth.service)
- auth/models.py: balance + credit_limit typed as Decimal (Numeric 15,2)
- pix/models.py: LedgerEntry.amount + PixTransaction fields -> Decimal
- boleto/models.py: value + fee_amount -> Decimal
- parcelamento/models.py: all float fields -> Decimal with proper precision
- minha_conta/models.py: plan_amount -> Decimal default 9.90
- pix/router.py: _build_pix_static_emv delegates to app.core.pix_emv canonical
- pix/internal_transfer.py: mask CPF/CNPJ and email in logs (PII)
- parcelamento/router.py: sanitized log (value+installments, no full body dump)
- web_routes.py: is_admin flag check replaces email comparison
- tests/test_auth_refresh.py: 11 tests covering silent rotation + rejection cases

Suite: 242 passed, 5 skipped, 0 failed

+++ app/boleto/models.py
@@ -1,6 +1,7 @@
-from sqlalchemy import Float, String, DateTime, Enum, Numeric
+from sqlalchemy import String, DateTime, Enum, Numeric
 from sqlalchemy.orm import Mapped, mapped_column
 from datetime import datetime, timezone
+from decimal import Decimal
 import enum
 from typing import Any, List
 from app.core.database import Base
@@ -21,7 +22,7 @@ class BoletoTransaction(Base):
     __tablename__ = "transacoes_boleto"
 
     id: Mapped[str] = mapped_column(String(36), primary_key=True, index=True)
-    value: Mapped[float] = mapped_column("valor", Numeric(15, 2, asdecimal=True), nullable=False)
+    value: Mapped[Decimal] = mapped_column("valor", Numeric(15, 2, asdecimal=True), nullable=False)
     barcode: Mapped[str] = mapped_column("codigo_barras", String(100), nullable=False)
     description: Mapped[str] = mapped_column("descricao", String(500), nullable=True)
     status: Mapped[BoletoStatus] = mapped_column(
@@ -33,4 +34,4 @@ class BoletoTransaction(Base):
     user_id: Mapped[str] = mapped_column(String(36), nullable=False, index=True)
     created_at: Mapped[datetime] = mapped_column("criado_em", DateTime, default=lambda: datetime.now(timezone.utc))
     correlation_id: Mapped[str] = mapped_column(String(100), nullable=True)
-    fee_amount: Mapped[float] = mapped_column("taxa_valor", Numeric(15, 2, asdecimal=True), nullable=True)
+    fee_amount: Mapped[Decimal] = mapped_column("taxa_valor", Numeric(15, 2, asdecimal=True), nullable=True)
